Update sidebar.html.eco
Just a quick PR to update something that is commonly confused: 'affect' vs 'effect'. It seems to me that was meant here was "to bring something about as a result". 

Before it read: "This should have no meaningful *affect* on your code"
How it should read: "This should have no meaningful *effect* on your code"